refactor: extract `Reporter`
